GetTickCount64 대신 QueryPerformanceCounter 를 사용하여 타이머의 정확도를 획기적으로 향상, 타이머 쓰기 작업을 Blocking Queue 를 사용한 생산자+소비자 패턴으로 작업 쓰레드를 분리하였음

+++ Direct3DHookLib/Interface/CaptureInterface.cs
@@ -253,7 +253,7 @@ namespace Direct3DHookLib.Interface
             SafeInvokeReportPropertyReceived(new ReportPropertyReceivedEventArgs(pid, dxVersion));
         }
 
-        public void ReportFps(float fps)
+        public void ReportFps(double fps)
         {
             SafeInvokeReportFpsReceived(new ReportFpsReceivedEventArgs(fps));
         }
